Add Prometheus for log-based replication
- Add Prometheus for log-based replication
