+++ src/server/utils/replication_utils.js
@@ -5,7 +5,18 @@ const _ = require('lodash');
 const dbg = require('../../util/debug_module')(__filename);
 const SensitiveString = require('../../util/sensitive_string');
 const system_store = require('../system_services/system_store').get_instance();
+const prom_reporting = require('../analytic_services/prometheus_reporting');
 const auth_server = require('../common_services/auth_server');
+
+const PARTIAL_SINGLE_BUCKET_REPLICATION_DEFAULTS = {
+    replication_id: '',
+    last_cycle_rule_id: '',
+    bucket_name: '',
+    last_cycle_writes_num: 0,
+    last_cycle_writes_size: 0,
+    last_cycle_error_writes_num: 0,
+    last_cycle_error_writes_size: 0,
+};
 
 //TODO: this function is not being used anymore, commenting out and keeping it as reference 
 // function check_data_or_md_changed(src_info, dst_info) {
@@ -28,6 +39,41 @@ const auth_server = require('../common_services/auth_server');
 //     return false;
 // }
 
+function get_rule_status(rule, src_cont_token, keys_diff_map, copy_res) {
+    const { num_keys_to_copy, num_bytes_to_copy } = Object.entries(keys_diff_map).reduce(
+        (acc, [key, value]) => {
+            acc.num_keys_to_copy += value.length;
+            acc.num_bytes_to_copy += value.reduce((key_bytes, obj) => key_bytes + obj.Size, 0);
+            return acc;
+        }, { num_keys_to_copy: 0, num_bytes_to_copy: 0 }
+    );
+
+    const num_keys_moved = copy_res.num_of_objects;
+    const num_bytes_moved = copy_res.size_of_objects;
+
+    const status = {
+        last_cycle_rule_id: rule,
+        last_cycle_writes_num: num_keys_moved,
+        last_cycle_writes_size: num_bytes_moved,
+        last_cycle_error_writes_num: num_keys_to_copy - num_keys_moved,
+        last_cycle_error_writes_size: num_bytes_to_copy - num_bytes_moved,
+    };
+    if (src_cont_token) status.last_cycle_src_cont_token = src_cont_token;
+    dbg.log1('get_rule_status: ', status);
+    return status;
+}
+
+function update_replication_prom_report(bucket_name, replication_policy_id, replication_status) {
+    const core_report = prom_reporting.get_core_report();
+    const last_cycle_status = _.defaults({
+        ...replication_status,
+        bucket_name: bucket_name.unwrap(),
+        replication_id: replication_policy_id
+    }, PARTIAL_SINGLE_BUCKET_REPLICATION_DEFAULTS);
+
+    core_report.set_replication_status(last_cycle_status);
+}
+
 /**
  * @param {any} bucket_name
  * @param {string} key
@@ -42,14 +88,16 @@ async function get_object_md(bucket_name, key, s3, version_id) {
         VersionId: version_id,
     };
 
+    dbg.log1('get_object_md params:', params);
     try {
         const head = await s3.headObject(params).promise();
         //for namespace s3 we are omitting the 'noobaa-namespace-s3-bucket' as it will be defer between buckets
         if (head?.Metadata) head.Metadata = _.omit(head.Metadata, 'noobaa-namespace-s3-bucket');
-        dbg.log1('BucketDiff _get_object_md: finished successfully', head);
+        dbg.log1('get_object_md: finished successfully', head);
         return head;
     } catch (err) {
-        dbg.error('BucketDiff _get_object_md: error:', err);
+        dbg.error('get_object_md: error:', err);
+        if (err.code === 'NotFound') return;
         throw err;
     }
 }
@@ -137,6 +185,8 @@ async function delete_objects(scanner_semaphore, client, bucket_name, keys) {
 }
 
 // EXPORTS
+exports.get_rule_status = get_rule_status;
+exports.update_replication_prom_report = update_replication_prom_report;
 exports.get_object_md = get_object_md;
 exports.find_src_and_dst_buckets = find_src_and_dst_buckets;
 exports.get_copy_type = get_copy_type;
